Add files via upload
For each item sold by the machine, the accept states for each automaton should be greater than or equal to the cost
of the item. 
The machine accepts $5, $10 and $20 coins,
The combination of coins to be returned as change is not important but there
The machine can only accept a maximum of $50 at any given moment.